requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
ultralytics==8.0.10
Pillow==10.0.0
torch==2.1.0
